+++ src/pages/BucketPage.js
@@ -1,6 +1,7 @@
 import { useRef } from "react";
 import styled from "styled-components";
 import { Coffee } from "../components/Coffee";
+import ScrollButton from "../components/ScrollButton";
 import useWindowScroll from "../hooks/useWindowScroll";
 
 const BucketWrapper = styled.div`
@@ -14,21 +15,12 @@ const BucketPage = () => {
     const scrollRef = useRef(null);
     const [hideElement, yHeight] = useWindowScroll(scrollRef);
 
-    const handleScroll = () => {
-        window.scrollBy({
-            top: yHeight,
-            behavior: 'smooth'
-        })
-    }
-
     return (
         <div>
             <section ref={scrollRef}>
                 <Coffee fadeIn={hideElement} />
             </section>
-            <div className="scroll-button">
-                <button onClick={handleScroll}>클릭</button>
-            </div>
+            <ScrollButton yCoord={yHeight} />
             <BucketWrapper>
                 본문
             </BucketWrapper>
